+++ templates/data/ship-classes.php
@@ -11,7 +11,7 @@ if(\is_array($shipClassList) && \count($shipClassList) > 0) {
 			<?php
 			foreach($shipClassList as $data) {
 				?>
-				<tr data-highlight="<?php echo $data['shipTypeSanitized']; ?>" onmouseover="dataHighlight('enable', '<?php echo $data['shipTypeSanitized']; ?>');" onmouseout="dataHighlight('disable', '<?php echo $data['shipTypeSanitized']; ?>');">
+				<tr data-highlight="shiptype-<?php echo $data['shipTypeSanitized']; ?>" onmouseover="dataHighlight('enable', 'shiptype-<?php echo $data['shipTypeSanitized']; ?>');" onmouseout="dataHighlight('disable', 'shiptype-<?php echo $data['shipTypeSanitized']; ?>');">
 					<td>
 						<?php
 						$image = \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\ImageHelper::getInstance()->getLocalCacheImageUriForRemoteImage('ship', \WordPress\Plugin\EveOnlineIntelTool\Libs\Helper\ImageHelper::getInstance()->getImageServerUrl('ship') . $data['shipID'] . '_32.png');