+++ templates/data/ship-types.php
@@ -11,7 +11,7 @@ if(\is_array($shipTypeList) && \count($shipTypeList) > 0) {
 			<?php
 			foreach($shipTypeList as $data) {
 				?>
-				<tr data-highlight="<?php echo $data['shipTypeSanitized']; ?>" onmouseover="dataHighlight('enable', '<?php echo $data['shipTypeSanitized']; ?>');" onmouseout="dataHighlight('disable', '<?php echo $data['shipTypeSanitized']; ?>');">
+				<tr data-highlight="shiptype-<?php echo $data['shipTypeSanitized']; ?>" onmouseover="dataHighlight('enable', 'shiptype-<?php echo $data['shipTypeSanitized']; ?>');" onmouseout="dataHighlight('disable', 'shiptype-<?php echo $data['shipTypeSanitized']; ?>');">
 					<td><?php echo $data['type']; ?></td>
 					<td class="table-data-count"><?php echo $data['count']; ?></td>
 				</tr>